backend/apps/buses/views/buses_views.py
from rest_framework import viewsets
from ..models import Bus
from ..serializers import BusSerializer


class BusViewSet(viewsets.ModelViewSet):
    queryset = Bus.objects.all()
    serializer_class = BusSerializer
